mkm: refactor to separate client state machine from epoll logic
