+++ components/mission_key_management/mkm.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <stdint.h>
+
+#define NUM_SECRET_KEYS 2
+#define SECRET_KEY_SIZE 32
+const uint8_t* get_mission_key(uint8_t key_id);
